pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "123@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-extensions (>=3.2.3,<4.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = false
strict_optional = true
no_implicit_optional = true
show_error_codes = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["django.*", "rest_framework.*", ".*.migrations.*"]
ignore_errors = false
disallow_untyped_defs = false


# [tool.mypy]
# strict_optional = true

[tool.django-stubs]
django_settings_module = "musicshop.settings"

# [tool.mypy.plugins]
# "mypy_django_plugin.main" = "django_stubs_ext"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
djangorestframework-stubs = { extras = [
    "compatible-mypy",
], version = "^3.15.3" }
django-stubs = { extras = ["compatible-mypy"], version = "^5.1.3" }
mypy = "^1.15.0"
types-requests = "^2.32.0.20250328"
mypy-extensions = "^1.0.0"
